+++ Rakefile
@@ -115,7 +115,6 @@ namespace :docs do
       Primer::CounterComponent,
       Primer::DetailsComponent,
       Primer::DropdownComponent,
-      Primer::DropdownMenuComponent,
       Primer::FlashComponent,
       Primer::HeadingComponent,
       Primer::LabelComponent,
