+++ app/components/primer/view_components.rb
@@ -36,7 +36,6 @@ require_relative "button_marketing_component"
 require_relative "counter_component"
 require_relative "details_component"
 require_relative "dropdown_component"
-require_relative "dropdown_menu_component"
 require_relative "flash_component"
 require_relative "flex_component"
 require_relative "flex_item_component"
